% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/auto_commit.R
\docType{methods}
\name{auto_commit}
\alias{auto_commit}
\alias{auto_commit,ANY,ANY-method}
\alias{auto_commit,ANY,git_connection-method}
\alias{auto_commit,git_connection-methods}
\title{Commit staged changes in a git repository with automated message}
\usage{
auto_commit(package, connection, username = character(0),
  password = character(0))

\S4method{auto_commit}{ANY,ANY}(package, connection, username = character(0),
  password = character(0))

\S4method{auto_commit}{ANY,git_connection}(package, connection,
  username = character(0), password = character(0))
}
\arguments{
\item{package}{The name of the package from which we autocommit}

\item{connection}{The path of the repository. Default to \code{rawdata.path}}

\item{username}{the username for the git repository}

\item{password}{the user's password for the git repository}
}
\description{
The mesagge is based on the information returned by \code{\link[utils]{sessionInfo}}
}

