% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/check_dataframe_covariate.R
\name{check_dataframe_covariate}
\alias{check_dataframe_covariate}
\title{Check if the covariates are available in a dataframe}
\usage{
check_dataframe_covariate(df, covariate, response = "Count", error = TRUE)
}
\arguments{
\item{df}{the \code{data.frame} to check}

\item{covariate}{The right hand side of the model as a character}

\item{response}{The left hand side of the model as a character}

\item{error}{When TRUE (default), the function returns an error when a variable is missing. Otherwise it returns a warning.}
}
\description{
Check if the covariates are available in a dataframe
}

