% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/check_dataframe_variable.R
\name{check_dataframe_variable}
\alias{check_dataframe_variable}
\title{Check if a data.frame contains variables}
\usage{
check_dataframe_variable(df, variable, name = "df", error = TRUE)
}
\arguments{
\item{df}{the \code{data.frame} to check}

\item{variable}{a vector with the names of the variable to check}

\item{name}{the name of the \code{data.frame} to use in the error message}

\item{error}{When TRUE (default), the function returns an error when a variable is missing. Otherwise it returns a warning.}
}
\value{
The function returns TRUE when all variables are present. If returns FALSE when a variable is missing and \code{error = FALSE}.
}
\description{
Check if a data.frame contains variables
}
\examples{
check_dataframe_variable(
 df = data.frame(a = integer(0)),
 variable = "a"
)
}

