% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/check_dbtable.R
\name{check_dbtable}
\alias{check_dbtable}
\title{Check if a table is available in a given ODBC connection}
\usage{
check_dbtable(table, schema = "dbo", channel, error = TRUE)
}
\arguments{
\item{table}{The name of the table}

\item{schema}{The schema. Defaults to 'dbo'}

\item{channel}{the open ODBC channel}

\item{error}{Indicates the behaviour when a table is missing. Gives an error when error = TRUE (default). Return FALSE otherwise.}
}
\value{
TRUE when all tables are present in the ODBC connection.
}
\description{
Check if a table is available in a given ODBC connection
}

