% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_dbtable_variable.R
\name{check_dbtable_variable}
\alias{check_dbtable_variable}
\title{Check if a variable is available in a given table}
\usage{
check_dbtable_variable(table, variable, schema = "public", channel,
  error = TRUE)
}
\arguments{
\item{table}{The name of the table}

\item{variable}{A vector with the names of the columns}

\item{schema}{The schema of the table. Defaults to public}

\item{channel}{the open dplyr connection to the database.}

\item{error}{Indicates the behaviour when a variable is missing. Gives an error when error = TRUE (default). Return FALSE otherwise.}
}
\value{
TRUE when all variables are present in the table.
}
\description{
Check if a variable is available in a given table
}
