% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_git_repo.R
\name{check_git_repo}
\alias{check_git_repo}
\title{Check if the path a git repository
Checks is a '.git' subdirectory exists in 'path'}
\usage{
check_git_repo(path, error = TRUE)
}
\arguments{
\item{path}{the path to check}

\item{error}{What to do in case \code{path} is not a git repository. Throw an
error when \code{error = TRUE}. Return \code{FALSE} when \code{error = FALSE}}
}
\description{
Check if the path a git repository
Checks is a '.git' subdirectory exists in 'path'
}

