% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/check_id.R
\name{check_id}
\alias{check_id}
\title{Test if an id exists in a given field of the table}
\usage{
check_id(value, field, table, channel)
}
\arguments{
\item{value}{the value of the id}

\item{field}{the field in which holds the id}

\item{table}{the table in which to look for the id}

\item{channel}{An open ODBC channel}
}
\description{
Test if an id exists in a given field of the table
}

