% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/check_path.R
\name{check_path}
\alias{check_path}
\title{check if a path is an exisiting file or directory}
\usage{
check_path(path, type = c("file", "directory"), error = TRUE)
}
\arguments{
\item{path}{the path of the directory or file name}

\item{type}{either "file" or "directory"}

\item{error}{When TRUE (default), the function returns an error when a variable is missing. Otherwise it returns a warning.}
}
\description{
check if a path is an exisiting file or directory
}

