% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_single_character.R
\name{check_single_character}
\alias{check_single_character}
\title{Check if the object is a single character}
\usage{
check_single_character(x, name = "x")
}
\arguments{
\item{x}{the object to check}

\item{name}{the name of the object to use in the error message}
}
\value{
The function gives the single character back. it throws an error when the input is not a single character.
}
\description{
Factors are converted to character.
}
\examples{
check_single_character("20")
}

