% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/check_single_logical.R
\name{check_single_logical}
\alias{check_single_logical}
\title{Check if the object is a single logical}
\usage{
check_single_logical(x, name = "x")
}
\arguments{
\item{x}{the object to check}

\item{name}{the name of the object to use in the error message}
}
\value{
The function gives the single logical back. it throws an error when the input is not a single logical.
}
\description{
Check if the object is a single logical
}
\examples{
check_single_logical(TRUE)
}

