% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_single_numeric.R
\name{check_single_numeric}
\alias{check_single_numeric}
\title{Check if the object is a single numeric}
\usage{
check_single_numeric(x, name = "x")
}
\arguments{
\item{x}{the object to check}

\item{name}{the name of the object to use in the error message}
}
\value{
The function gives the single numeric back. it throws an error when the input is not a single numeric.
}
\description{
Check if the object is a single numeric
}
\examples{
check_single_numeric(0.5)
}

