% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/check_single_posix.R
\name{check_single_posix}
\alias{check_single_posix}
\title{Check if the object is a single POSIX}
\usage{
check_single_posix(x, name = "x", past = FALSE)
}
\arguments{
\item{x}{the object to check}

\item{name}{the name of the object to use in the error message}

\item{past}{Should the function throw an error when x is in the future? Default is FALSE.}
}
\value{
The function gives the single POSIX back. It throws an error when the input is not a single character.
}
\description{
Check if the object is a single POSIX
}
\examples{
check_single_posix(Sys.time())
}

