% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_single_probability.R
\name{check_single_probability}
\alias{check_single_probability}
\title{Check if the object is a single probability}
\usage{
check_single_probability(x, name = "x")
}
\arguments{
\item{x}{the object to check}

\item{name}{the name of the object to use in the error message}
}
\value{
The function gives the single probability back. it throws an error when the input is not a single probability.
}
\description{
Check if the object is a single probability
}
\examples{
check_single_probability(0.5)
}
