% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/check_single_strictly_positive_integer.R
\name{check_single_strictly_positive_integer}
\alias{check_single_strictly_positive_integer}
\title{Check if the object is a single strictly positive integer}
\usage{
check_single_strictly_positive_integer(x, name = "x", tolerance = 1e-10)
}
\arguments{
\item{x}{the object to check}

\item{name}{the name of the object to use in the error message}

\item{tolerance}{was it an acceptable difference in case x is numeric}
}
\value{
The function the strictly positive integer back. it throws an error when the input is not a strictly positive integer
}
\description{
Check if the object is a single strictly positive integer
}
\examples{
check_single_strictly_positive_integer(20)
}

