% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect_result.R
\name{connect_ut_db}
\alias{connect_ut_db}
\title{connect to the unit test database}
\usage{
connect_ut_db(host = "localhost", dbname = "n2kunittest",
  user = "unittest_analysis", password = "unittest", port = 5432,
  ...)
}
\arguments{
\item{host}{Host name and port number of database}

\item{dbname}{Database name}

\item{user}{User name and password.

Generally, you should avoid saving username and password in your
scripts as it is easy to accidentally expose valuable credentials.
Instead, retrieve them from environment variables, or use database
specific credential scores. For example, with MySQL you can set up \code{my.cnf}
as described in \code{\link[RMySQL:MySQL]{RMySQL::MySQL()}}.}

\item{password}{User name and password.

Generally, you should avoid saving username and password in your
scripts as it is easy to accidentally expose valuable credentials.
Instead, retrieve them from environment variables, or use database
specific credential scores. For example, with MySQL you can set up \code{my.cnf}
as described in \code{\link[RMySQL:MySQL]{RMySQL::MySQL()}}.}

\item{port}{Host name and port number of database}

\item{...}{for the src, other arguments passed on to the underlying
database connector, \code{\link[DBI:dbConnect]{DBI::dbConnect()}}. For the tbl, included for
compatibility with the generic, but otherwise ignored.}
}
\description{
connect to the unit test database
}
