% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/get_datasource_id.R
\name{get_datasource_id}
\alias{get_datasource_id}
\title{Get the id of a data source}
\usage{
get_datasource_id(data.source.name, develop = TRUE)
}
\arguments{
\item{data.source.name}{the description of the data source}

\item{develop}{Logical value. Indicates the location ton the results database}
}
\description{
Get the id of a data source
}

