% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/get_scheme_id.R
\name{get_scheme_id}
\alias{get_scheme_id}
\title{Get the id of a scheme}
\usage{
get_scheme_id(scheme.name, develop = TRUE)
}
\arguments{
\item{scheme.name}{the name of the scheme}

\item{develop}{Logical value. Indicates the location ton the results database}
}
\description{
Get the id of a scheme
}

