% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/git_sha.R
\docType{methods}
\name{git_sha}
\alias{git_sha}
\alias{git_sha,ANY,ANY-method}
\alias{git_sha,ANY,git_connection-method}
\alias{git_sha,git_connection-methods}
\title{Get the SHA of the files at the HEAD}
\usage{
git_sha(file, connection, path)

\S4method{git_sha}{ANY,ANY}(file, connection, path)

\S4method{git_sha}{ANY,git_connection}(file, connection, path)
}
\arguments{
\item{file}{the name of the file}

\item{connection}{The path of a git repository or a \code{git_connection} object}

\item{path}{the subdirectory within the git repository. Ignored when \code{connection} is a \code{git connection} object.}
}
\description{
Get the SHA of the files at the HEAD
}

