% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_chartor.R
\name{is_chartor}
\alias{is_chartor}
\title{Test if the argument is either character or factor}
\usage{
is_chartor(x)
}
\arguments{
\item{x}{the object to check}
}
\description{
Test if the argument is either character or factor
}
