% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/is_git_repo.R
\name{is_git_repo}
\alias{is_git_repo}
\title{It the path a git repository
Checks is a '.git' subdirectory exists in 'path'}
\usage{
is_git_repo(path)
}
\arguments{
\item{path}{the path to check}
}
\value{
A logical vector wit the same length as path
}
\description{
It the path a git repository
Checks is a '.git' subdirectory exists in 'path'
}

