% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/list_files_git.R
\name{list_files_git}
\alias{list_files_git}
\title{List the files in a path of a git repository}
\usage{
list_files_git(path, pattern = NULL, repo.path, full.names = FALSE)
}
\arguments{
\item{path}{the subdirectory of the file relative to the root of the git repository}

\item{pattern}{an optional \link{regular expression}.  Only file names
    which match the regular expression will be returned.}

\item{repo.path}{path of the git repository. Defaults to rawdata.path}

\item{full.names}{a logical value.  If \code{TRUE}, the directory
    path is prepended to the file names to give a relative file path.
    If \code{FALSE}, the file names (rather than paths) are returned.}
}
\description{
List the files in a path of a git repository
}

