% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/match_nbn_key.R
\name{match_nbn_key}
\alias{match_nbn_key}
\title{Merge NBN keys into a species dataframe}
\usage{
match_nbn_key(species, nbn.key, variable)
}
\arguments{
\item{species}{a data.frame with the species names}

\item{nbn.key}{a data.frame with the NBN keys}

\item{variable}{the name of the variable of species to match with InputName
from nbn.key}
}
\description{
Merge NBN keys into a species dataframe
}

