% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/odbc_connect.R
\name{odbc_connect}
\alias{odbc_connect}
\title{connect to a data source through ODBC}
\usage{
odbc_connect(data.source.name, develop = TRUE)
}
\arguments{
\item{data.source.name}{The name of the data source}

\item{develop}{Logical value. Indicates the location ton the results database}
}
\description{
The connection string is stored in the results database.
}

