% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/odbc_get_id.R
\name{odbc_get_id}
\alias{odbc_get_id}
\title{Get the id of the mathcing records}
\usage{
odbc_get_id(table, variable, value, develop = TRUE)
}
\arguments{
\item{table}{the name of the}

\item{variable}{the name of the variable}

\item{value}{the content of the variable}

\item{develop}{Logical value. Indicates the location ton the results database}
}
\description{
Get the id of the mathcing records
}

