% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/read_delim_git.R
\name{read_delim_git}
\alias{read_delim_git}
\title{read a tab delimited file from a git repository}
\usage{
read_delim_git(file, path, repo.path = rawdata.path)
}
\arguments{
\item{file}{the name of the file}

\item{path}{the subdirectory of the file relative to the root of the git repository}

\item{repo.path}{path of the git repository. Defaults to rawdata.path}
}
\description{
read a tab delimited file from a git repository
}

