% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_object_environment.R
\name{read_object_environment}
\alias{read_object_environment}
\title{Read an object from an environment}
\usage{
read_object_environment(object, env, warn = TRUE)
}
\arguments{
\item{object}{the name of the object}

\item{env}{the environment}

\item{warn}{Issue a warning if the object is not found in the environment.
Defaults to TRUE}
}
\value{
the object or \code{NULL} is the object doesn't exists in the
environment
}
\description{
Read an object from an environment
}
\examples{
  object <- "test"
  value <- TRUE
  env <- new.env()
  assign(x = object, value = value, envir = env)
  read_object_environment(object, env)
}
