% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/remove_files_git.R
\name{remove_files_git}
\alias{remove_files_git}
\title{Remove all the files in a path of a git repository}
\usage{
remove_files_git(path, pattern = NULL, repo.path)
}
\arguments{
\item{path}{the subdirectory of the file relative to the root of the git repository}

\item{pattern}{an optional \link{regular expression}.  Only file names
    which match the regular expression will be returned.}

\item{repo.path}{path of the git repository. Defaults to rawdata.path}
}
\description{
Note that the removal is not staged
}

