% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_sha1.R
\name{sha1_digits}
\alias{sha1_digits}
\title{Define the number of digits to use when rounding numeric values}
\usage{
sha1_digits(which = c("base", "zapsmall", "coef"))
}
\arguments{
\item{which}{The number of digits for each usage}
}
\description{
Define the number of digits to use when rounding numeric values
}

