% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/write_delim_git.R
\name{write_delim_git}
\alias{write_delim_git}
\title{write a dataframe as a tab delimited file to a git repository and stage it
The existing file will be overwritten.}
\usage{
write_delim_git(x, file, path, repo.path = rawdata.path)
}
\arguments{
\item{x}{the data.frame}

\item{file}{the name of the file}

\item{path}{the subdirectory of the file relative to the root of the git repository}

\item{repo.path}{path of the git repository. Defaults to rawdata.path}
}
\description{
write a dataframe as a tab delimited file to a git repository and stage it
The existing file will be overwritten.
}

