% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/get_sha1.R
\name{zap_small}
\alias{zap_small}
\title{Replace tiny numbers with zero}
\usage{
zap_small(x, digits = sha1_digits("zapsmall"))
}
\arguments{
\item{x}{the numeric vector to check}

\item{digits}{The minimal magnitude $10^-{digits}$}
}
\description{
Replace tiny numbers with zero
}

