% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_dataframe_variable.R
\name{check_dataframe_variable}
\alias{check_dataframe_variable}
\title{Check if a data.frame contains variables}
\usage{
check_dataframe_variable(
  df,
  variable,
  name = "df",
  force_na = FALSE,
  error = TRUE
)
}
\arguments{
\item{df}{the \code{data.frame} to check}

\item{variable}{either a character vector with the names of the variable to
check or a named list. The names of the list must match the names of the
required variables in the data.frame. The elements of the list contain the
accepted classes for each varaible.}

\item{name}{the name of the \code{data.frame} to use in the error message}

\item{force_na}{check the class of variables with all NA}

\item{error}{When TRUE (default), the function returns an error when a
variable is missing. Otherwise it returns a warning.}
}
\value{
The function returns TRUE when all variables are present. If returns
FALSE when a variable is missing and \code{error = FALSE}.
}
\description{
Check if a data.frame contains variables
}
\examples{
check_dataframe_variable(
 df = data.frame(a = integer(0)),
 variable = "a"
)
check_dataframe_variable(
 df = data.frame(a = integer(0)),
 variable = list(a = c("integer", "numeric"))
)
}
