% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_id.R
\name{check_id}
\alias{check_id}
\title{Test if an id exists in a given field of the table}
\usage{
check_id(value, variable, table, channel)
}
\arguments{
\item{value}{the content of the variable}

\item{variable}{A vector with the names of the columns}

\item{table}{The name of the table}

\item{channel}{the open dplyr connection to the database.}
}
\description{
Test if an id exists in a given field of the table
}
