% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect_result.R
\name{connect_result}
\alias{connect_result}
\title{Opens an ODBC connection to the 'results' database}
\usage{
connect_result(username, password, develop = TRUE)
}
\arguments{
\item{username}{the username to connect to the database.}

\item{password}{the password for the username.}

\item{develop}{Logical value. Indicates the location of the results database}
}
\description{
Opens an ODBC connection to the 'results' database
}
