% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/connect_result.R
\name{connect_result}
\alias{connect_result}
\title{Opens an ODBC connection to the 'results' database}
\usage{
connect_result(develop = TRUE)
}
\arguments{
\item{develop}{Logical value. Indicates the location ton the results database}
}
\description{
Opens an ODBC connection to the 'results' database
}

