% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect_result.R
\name{connect_ut_db}
\alias{connect_ut_db}
\title{connect to the unit test database}
\usage{
connect_ut_db(
  host = "localhost",
  dbname = "n2kunittest",
  user = "unittest_analysis",
  password = "unittest",
  port = 5432,
  ...
)
}
\arguments{
\item{host}{Host name and port number of PostgreSQL database.}

\item{dbname}{Database name.}

\item{user}{User name and password.}

\item{password}{User name and password.}

\item{port}{Port number of database.
Defaults to 5432}

\item{...}{arguments past to \code{\link[DBI:dbConnect]{DBI::dbConnect()}}.}
}
\description{
connect to the unit test database
}
