% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/get_nbn_key.R
\name{get_nbn_key}
\alias{get_nbn_key}
\title{Get the NBN key of a species}
\usage{
get_nbn_key(name, language = "la", develop = TRUE)
}
\arguments{
\item{name}{a vector of species names to check}

\item{language}{The language to use. Defaults to "la" 'scientific name"}

\item{develop}{Logical value. Indicates the location ton the results database}
}
\description{
Get the NBN key of a species
}

