% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_nbn_key_multi.R
\name{get_nbn_key_multi}
\alias{get_nbn_key_multi}
\title{Try multiple languages to get a matching NBN key}
\usage{
get_nbn_key_multi(species, orders = c("la", "nl", "en"), channel)
}
\arguments{
\item{species}{A data.frame with the name of species in one or more languages}

\item{orders}{the order in which the languages are tried to get a matching
NBN key.}

\item{channel}{An open RODBC channel to the NBN database}
}
\description{
Try multiple languages to get a matching NBN key
}
