% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_nbn_name.R
\name{get_nbn_name}
\alias{get_nbn_name}
\title{Get the name associated with an NBN key}
\usage{
get_nbn_name(nbn_key, channel)
}
\arguments{
\item{channel}{An open RODBC channel to the NBN database}

\item{nbn.key}{A vector with NBN keys}
}
\description{
Get the name associated with an NBN key
}
