% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/git_connect.R
\name{git_connect}
\alias{git_connect}
\title{returns the path of the datasource within the git repository}
\usage{
git_connect(data.source.name, develop = TRUE)
}
\arguments{
\item{data.source.name}{The name of the data source}

\item{develop}{Logical value. Indicates the location ton the results database}
}
\description{
The details are stored in the results database.
}

