% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/git_sha.R
\name{git_sha}
\alias{git_sha}
\title{Get the SHA of the files at the HEAD}
\usage{
git_sha(file, path, repo.path = rawdata.path)
}
\arguments{
\item{file}{the name of the files}

\item{path}{the subdirectory of the file relative to the root of the git repository}

\item{repo.path}{path of the git repository. Defaults to rawdata.path}
}
\description{
Get the SHA of the files at the HEAD
}

