% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/odbc_connect.R
\name{odbc_connect}
\alias{odbc_connect}
\title{connect to a data source through ODBC}
\usage{
odbc_connect(data_source_name, username, password, channel)
}
\arguments{
\item{data_source_name}{The name of the data source}

\item{username}{the username in case the ConnectMethod is \code{"Credentials supplied by the user running the report"}.
Ignored in all other cases.}

\item{password}{the password to be used in combination with the username.}

\item{channel}{the ODBC channel to the database with the connection strings}
}
\description{
The connection string is stored in the results database.
}
