% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/odbc_get_id.R
\name{odbc_get_id}
\alias{odbc_get_id}
\title{Get the id of the matching records}
\usage{
odbc_get_id(table, variable, value, schema = "dbo", channel)
}
\arguments{
\item{table}{The name of the table}

\item{variable}{A vector with the names of the columns}

\item{value}{the content of the variable}

\item{schema}{The schema of the table. Defaults to dbo}

\item{channel}{the open ODBC channel}
}
\description{
Get the id of the matching records
}

