% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/odbc_get_id.R
\name{odbc_get_id}
\alias{odbc_get_id}
\title{Get the corresponding id's}
\usage{
odbc_get_id(data, id.field, merge.field, table, channel, create = FALSE)
}
\arguments{
\item{data}{the data.frame}

\item{id.field}{the id fields}

\item{merge.field}{the merge fields}

\item{table}{the table name}

\item{channel}{the open ODBC channel}

\item{create}{When TRUE, the function creates unmatching records AND updates attributes. Defaults to FALSE.}
}
\value{
a data.frame with data and the id's
}
\description{
Get the corresponding id's
}

