% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/odbc_insert.R
\name{odbc_insert}
\alias{odbc_insert}
\title{Append a data.frame to a table through an ODBC connection}
\usage{
odbc_insert(data, table, channel, schema = "dbo", append = TRUE,
  rows.at.time = 1000)
}
\arguments{
\item{data}{the data.frame}

\item{table}{The name of the table}

\item{channel}{the open dplyr connection to the database.}

\item{schema}{The schema of the table. Defaults to public}

\item{append}{Append the data or overwrite existing rows?}

\item{rows.at.time}{Number of rows to insert in one SQL statement}
}
\value{
The status of the SQL INSERT for each row in returned but invisible.
}
\description{
Append a data.frame to a table through an ODBC connection
}
