% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/odbc_insert.R
\name{odbc_insert}
\alias{odbc_insert}
\title{Append a data.frame to a table through an ODBC connection}
\usage{
odbc_insert(data, table, channel, schema = "dbo", append = TRUE)
}
\arguments{
\item{data}{the data.frame}

\item{table}{The name of the table}

\item{channel}{the open ODBC channel}

\item{schema}{The schema of the table. Defaults to dbo}

\item{append}{Append the data or overwrite existing rows?}
}
\value{
The status of the SQL INSERT for each row in returned but invisible.
}
\description{
Append a data.frame to a table through an ODBC connection
}

