% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/odbc_insert.R
\name{odbc_insert}
\alias{odbc_insert}
\title{Append a data.frame to a table through an ODBC connection}
\usage{
odbc_insert(channel, data, table)
}
\arguments{
\item{channel}{An open ODBC channel}

\item{data}{the data.frame}

\item{table}{the table in which to look for the id}
}
\value{
The status of the SQL INSERT for each row in returned but invisible.
}
\description{
Append a data.frame to a table through an ODBC connection
}

