% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/read_delim_git.R
\docType{methods}
\name{read_delim_git}
\alias{read_delim_git}
\alias{read_delim_git,ANY,ANY-method}
\alias{read_delim_git,ANY,git_connection-method}
\alias{read_delim_git,git_connection-methods}
\title{Read a tab delimited file from a git repository}
\usage{
read_delim_git(file, connection, path)

\S4method{read_delim_git}{ANY,ANY}(file, connection, path)

\S4method{read_delim_git}{ANY,git_connection}(file, connection, path)
}
\arguments{
\item{file}{the name of the file}

\item{connection}{The path of a git repository or a \code{git_connection} object}

\item{path}{the subdirectory within the git repository. Ignored when \code{connection} is a \code{git connection} object.}
}
\description{
Read a tab delimited file from a git repository
}

