% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/remove_files_git.R
\docType{methods}
\name{remove_files_git}
\alias{remove_files_git}
\alias{remove_files_git,ANY-method}
\alias{remove_files_git,git_connection-method}
\alias{remove_files_git,git_connection-methods}
\title{Remove all the files in a path of a git repository}
\usage{
remove_files_git(connection, path, pattern = NULL)

\S4method{remove_files_git}{ANY}(connection, path, pattern = NULL)

\S4method{remove_files_git}{git_connection}(connection, path, pattern = NULL)
}
\arguments{
\item{connection}{The path of a git repository or a \code{git_connection} object}

\item{path}{the subdirectory within the git repository. Ignored when \code{connection} is a \code{git connection} object.}

\item{pattern}{an optional \link{regular expression}.  Only file names
    which match the regular expression will be returned.}
}
\description{
Remove all the files in a path of a git repository
}

