% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/write_delim_git.R
\docType{methods}
\name{write_delim_git}
\alias{write_delim_git}
\alias{write_delim_git,ANY,ANY,ANY-method}
\alias{write_delim_git,ANY,ANY,git_connection-method}
\alias{write_delim_git,git_connection-methods}
\title{Write a dataframe as a tab delimited file to a git repository and stage it}
\usage{
write_delim_git(x, file, connection, path)

\S4method{write_delim_git}{ANY,ANY,ANY}(x, file, connection, path)

\S4method{write_delim_git}{ANY,ANY,git_connection}(x, file, connection, path)
}
\arguments{
\item{x}{the data.frame}

\item{file}{the name of the file}

\item{connection}{The path of a git repository or a \code{git_connection} object}

\item{path}{the subdirectory within the git repository. Ignored when \code{connection} is a \code{git connection} object.}
}
\value{
the SHA1 of the file
}
\description{
The existing file will be overwritten.
}

